pyproject.toml
[project]
name = "dbt-cll-evaluator"
version = "0.1.0"
description = "CLI to leverage insights from your Column Level Lineage"
readme = "README.md"
dependencies = [
    "typer>=0.12.4",
    "networkx>=3.3",
    "matplotlib>=3.9.2",
    "httpx>=0.27.0",
]
requires-python = ">=3.10"

[tool.uv]
dev-dependencies = [
    "ruff >= 0.5.2",
]

[tool.poetry]
name = "dbt-cll-evaluator"
version = "0.1.0"
description = "CLI to leverage insights from your Column Level Lineage"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
typer = "^0.12.3"
httpx = "^0.27.0"
networkx = "^3.3"
matplotlib = "^3.9.1"
pydantic = "^2.9.1"
terminaltexteffects = "^0.11.0"
ruamel-yaml = "^0.18.6"
loguru = "^0.7.2"

[tool.poetry.group.dev.dependencies]
ruff = "^0.5.2"

[tool.poetry.scripts]
dbt-cll-evaluator = "dbt_cll_evaluator.__main__:startup"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
